crawler-backend/requirements.txt
pymongo~=4.6.2
fastapi~=0.109.2
pydantic~=1.10.11
beanie~=1.25.0
motor~=3.3.2
python-dotenv~=1.0.1
uvicorn~=0.27.1